Changed parameter types in ICollidableC:s getBoundingBox and changed pararmeters in the same method in ICollidable

+++ project/HowDoIFlyThisThing/src/main/java/com/whathappensingandalf/howdoiflythisthing/ICollidableComponent.java
@@ -9,6 +9,6 @@ import java.awt.geom.Rectangle2D;
 public interface ICollidableComponent {
 	
 	public void updateBoundingBox(double x, double y, double width, double height);
-	public Rectangle2D getBoundingBox(float x, float y, double width, double height);
+	public Rectangle2D getBoundingBox(double x, double y, double width, double height);
 	public boolean collideDetection(ICollidable rhs, double x, double y, double width, double height);
 }
